pyproject.toml
[project]
name = "batchblast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.121.0",
    "perplexityai>=0.20.0",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
    "wsproto>=1.2.0",
]
